+++ plugins/11_TheWarWithin/common.lua
@@ -148,6 +148,12 @@ ns.groups.YOU_CANT_HANG_WITH_US = Group('you_cant_hang_with_us', 5763494, {
     achievement = 40634
 })
 
+ns.groups.CRITTER_LOVE = Group('critter_love', 3459801, {
+    defaults = ns.GROUP_HIDDEN,
+    type = ns.group_types.ACHIEVEMENT,
+    achievement = 40475
+})
+
 -------------------------------------------------------------------------------
 ---------------------------- KHAZ ALGAR LORE HUNTER ---------------------------
 -------------------------------------------------------------------------------