+++ plugins/11_TheWarWithin/localization/esES.lua
@@ -99,6 +99,10 @@ L['options_icons_gobblin_with_glublurp'] = '{achievement:40614}'
 L['options_icons_gobblin_with_glublurp_desc'] = 'Muestra los lugares para el logro {achievement:40614}.'
 L['gobblin_with_glublurp_note'] = 'Haz Click en un {dot:Red}{object:Cristal Trémulo} para ganar {spell:456739}. Coge una {npc:227138} volando en el ' .. ns.color.Orange('Círculo Naranja') .. ' y llévasela a {npc:227132}.\n\n(Se recomienda Vuelo Constante para conseguilo).'
 
+L['critter_love_note'] = nil
+L['options_icons_critter_love'] = nil
+L['options_icons_critter_love_desc'] = nil
+
 -------------------------------------------------------------------------------
 --------------------------------- HALLOWFALL ----------------------------------
 -------------------------------------------------------------------------------